requirements.txt
fastapi
uvicorn
sqlalchemy
databases
mysql-connector-python
passlib
bcrypt
python-jose
python-multipart
user-agents
geoip2
jwt
aiohttp
aiosignal
aiosmtplib
annotated-types
anyio
attrs
blinker
certifi
pillow
pandas
